% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alkalinity.R
\name{alkalinity}
\alias{alkalinity}
\title{Calculates solution alkalinity}
\usage{
alkalinity(
  pH,
  nDIC,
  VolumeSolution,
  VolumeHeadspace,
  temperature,
  K.CO2HCO3,
  K.HCO3CO3
)
}
\arguments{
\item{nDIC}{Moles of dissolved inorganic carbon}

\item{VolumeSolution}{Volume of liquid in the closed system, in liters.}

\item{VolumeHeadspace}{Volume of gaseous headspace in the closed system, in liters.}

\item{temperature}{Temperature of the system, in Kelvin.}

\item{K.CO2HCO3}{Equilibrium constant for the dissociation of CO2(aq) to HCO3-(aq).}

\item{pH.initial}{Initial pH}

\item{KHCO3CO3}{Equilibrium constant for the dissociation of HCO3- (aq) to CO3-- (aq).}
}
\value{
alkalinity of the solution, in millimolar.
}
\description{
`alkalinity()` Uses the calculate_closed_system_alkalinity function from Sebastian Kopf's microbialkitchen R package to calculate media alkalinity
}
