% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqueous.R
\name{aqueous.step}
\alias{aqueous.step}
\title{Uses Henry's Law to calculate dissolved gas concentrations}
\usage{
aqueous.step(P.step, K)
}
\description{
`aqueous.step()` calculates the aqueous concentration at each step using the current pressure.
}
\details{
@param P.step the current pressure in the system
 @param K Henry's equilibrium constant of the gas
}
