% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{biomass.coefficient}
\alias{biomass.coefficient}
\title{Amount of biomass produced for every DIC consumed}
\usage{
biomass.coefficient(biomass.yield, carbon.fraction)
}
\arguments{
\item{biomass.yield}{mass of dry biomass produced per mol of product.}

\item{carbon.fraction}{w/w percent C of biomass, expressed as a decimal.}
}
\value{
biomass produced, in mol/mol DIC
}
\description{
`biomass.coefficient()` calculates the amount of biomass produced for every mol of DIC consumed
}
