% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.K.R
\name{calculate.Kgas}
\alias{calculate.Kgas}
\title{Calculates Henry's constant for a gas}
\usage{
calculate.Kgas(bunsen, molar.vol.gas)
}
\arguments{
\item{bunsen}{The Bunsen coefficient for the gas.}

\item{molar.vol.gas}{The moles per unit volume of gas.}
}
\description{
`calculate.Kgas()` uses the provided Bunsen coefficient and moles per volume of gas to calculate Henry's constant.
}
