% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{carbon.yield}
\alias{carbon.yield}
\title{Biomass to carbon yield}
\usage{
carbon.yield(biomass.yield, carbon.fraction)
}
\arguments{
\item{biomass.yield}{mass of dry biomass produced per mol of product.}

\item{carbon.fraction}{w/w percent C of biomass, expressed as a decimal.}
}
\value{
carbon yield, in mol C/mol product
}
\description{
`carbon.yield()` converts the known organism-specific biomass yield per mol of product into carbon yield.
}
