% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{CH4.coefficient}
\alias{CH4.coefficient}
\title{Amount of CH4 produced for every DIC consumed}
\usage{
CH4.coefficient(biomass.yield, carbon.fraction)
}
\arguments{
\item{biomass.yield}{mass of dry biomass produced per mol of product.}

\item{carbon.fraction}{w/w percent C of biomass, expressed as a decimal.}
}
\value{
CH4 produced, in mol/mol DIC
}
\description{
`CH4.coefficient()` calculates the amount of CH4 produced for every mol of DIC consumed.
}
