% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs.R
\name{gibbs.step}
\alias{gibbs.step}
\title{Calculates Gibbs free energy}
\usage{
gibbs.step(standard.gibbs, Q, temperature)
}
\arguments{
\item{standard.gibbs}{The standard Gibbs free energy of formation of the reaction in J/mol}

\item{Q}{Current reaction quotient}

\item{temperature}{Temperature of the system, in Kelvin.}
}
\description{
`gibbs.step()` calculates the Gibbs free energy at each step in the
model for a given Q and standard Gibbs free energy.
}
