% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{H2.coefficient}
\alias{H2.coefficient}
\title{Amount of H2 consumed for every DIC consumed}
\usage{
H2.coefficient(biomass.yield, carbon.fraction)
}
\arguments{
\item{biomass.yield}{mass of dry biomass produced per mol of product. 2.4 g/mol product by default.}

\item{carbon.fraction}{w/w percent C of biomass, expressed as a decimal. 0.44 by default.}
}
\value{
H2 consumed, in mol/mol DIC
}
\description{
`H2.coefficient()` calculates the amount of H2 consumed for every mol of DIC consumed.
}
