% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methanogenesis.R
\name{methanogenesis.time}
\alias{methanogenesis.time}
\title{Steps through time during a methanogenesis reaction}
\usage{
methanogenesis.time(
  CH4.initial,
  K.CH4 = NA,
  H2.initial,
  K.H2 = NA,
  is.H2.limiting = NA,
  Ks.H2,
  DIC.initial,
  pH.initial,
  K.CO2 = NA,
  is.CO2.limiting = NA,
  Ks.CO2,
  umax,
  temperature,
  VolumeSolution,
  VolumeHeadspace,
  K.CO2HCO3 = NA,
  K.HCO3CO3 = NA,
  time.step = 0.1,
  total.time = 5,
  inoculum.cell.number = 1e+06,
  biomass.yield = 2.4,
  carbon.fraction = 0.44,
  cell.weight = 3e-14
)
}
\arguments{
\item{CH4.initial}{Concentration of initial dissolved CH4, in molarity.}

\item{K.CH4}{Henry's constant for CH4. NA by default (calculated by CHNOSZ).}

\item{H2.initial}{Concentration of initial dissolved H2, in molarity.}

\item{K.H2}{Henry's constant for H2. NA by default (calculated by CHNOSZ).}

\item{is.H2.limiting}{Boolean. Determines whether to use the Monod equation for H2. NA by default (limiting contribution determined by the proportion of current H2 and Ks.H2).}

\item{Ks.H2}{Half-saturation constant for H2 (microorganism dependent).}

\item{DIC.initial}{Concentration of initial dissolved inorganic carbon, in molarity.}

\item{pH.initial}{initial pH.}

\item{K.CO2}{Henry's constant for CH4. NA by default (calculated by CHNOSZ).}

\item{is.CO2.limiting}{Boolean. Determines whether to use the Monod equation for CO2. NA by default (limiting contribution determined by the proportion of current CO2 and Ks.CO2).}

\item{Ks.CO2}{Half-saturation constant for CO2 (microorganism dependent).}

\item{umax}{Maximum growth rate for the microorganism, \ifelse{html}{\out{hr<sup>-1</sup>}}{\eqn{hr^-1}}}

\item{temperature}{Temperature of the system, in Kelvin.}

\item{VolumeSolution}{Volume of liquid in the closed system, in liters.}

\item{VolumeHeadspace}{Volume of gaseous head space in the closed system, in liters.}

\item{K.CO2HCO3}{Equilibrium constant for the dissociation of CO2(aq) to HCO3-(aq). NA by default (calculated by CHNOSZ).}

\item{K.HCO3CO3}{Equilibrium constant for the dissociation of HCO3- (aq) to CO3-- (aq). NA by default (calculated by CHNOSZ).}

\item{time.step}{Step size in time, in hours. 0.1 hrs by default.}

\item{total.time}{Length of time the model will run, in hours. Note: the model may break before the total.time is reached if H2 or DIC runs out.}

\item{inoculum.cell.number}{Initial number of cells. 1e6 by default.}

\item{biomass.yield}{mass of dry biomass produced per mol of product. 2.4 g/mol product by default.}

\item{carbon.fraction}{w/w percent C of biomass, expressed as a decimal. 0.44 by default.}

\item{cell.weight}{individual cell mass, in grams. 30e-15 by default.}
}
\value{
A data frame of the model results
}
\description{
`methanogenesis.time()` calculates CH4 produced, H2 consumed, CO2 consumed, pH, and Gibbs free energy changes as dissolved inorganic carbon is consumed over time.
}
\examples{
methanogenesis(CH4.initial = 1e-6,H2.initial = 5e-4,DIC.initial = 3.2e-3,pH.initial = 7.5,temperature = 273.15+40,VolumeSolution = 80e-3,VolumeHeadspace = 20e-3,delta.DIC = 0.0001)

}
