% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pH.R
\name{pH}
\alias{pH}
\title{Calculates pH of the solution}
\usage{
pH(
  nDIC,
  VolumeSolution,
  VolumeHeadspace,
  temperature,
  alkalinity,
  K.CO2HCO3,
  K.HCO3CO3
)
}
\arguments{
\item{nDIC}{Moles of dissolved inorganic carbon}

\item{VolumeSolution}{Volume of liquid in the closed system, in liters.}

\item{VolumeHeadspace}{Volume of gaseous head space in the closed system, in liters.}

\item{temperature}{Temperature of the system, in Kelvin.}

\item{alkalinity}{initial alkalinity of the system, in millimolar.}

\item{K.CO2HCO3}{Equilibrium constant for the dissociation of CO2(aq) to HCO3-(aq).}

\item{K.HCO3CO3}{Equilibrium constant for the dissociation of HCO3- (aq) to CO3– (aq).}
}
\value{
pH of the solution.
}
\description{
`pH()` Uses the calculate_closed_system_pH function from Sebastian Kopf's microbialkitchen R package to calculate pH.
}
