% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressure.R
\name{pressure.step}
\alias{pressure.step}
\title{Calculates head space pressure}
\usage{
pressure.step(n.total.step, K, VolumeSolution, VolumeHeadspace, temperature)
}
\arguments{
\item{n.total.step}{The total number of moles of a gas in the system.}

\item{K}{Henry's constant of the gas.}

\item{VolumeSolution}{Volume of liquid in the closed system, in liters.}

\item{VolumeHeadspace}{Volume of gaseous head space in the closed system, in liters.}

\item{temperature}{Temperature of the system, in Kelvin.}
}
\value{
Partial pressure of the head space.
}
\description{
`pressure.step()` calculates head space partial pressure from the total moles of a particular gas in the system using Henry's law and the Ideal Gas law.
}
