% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methanogenesis.R
\name{methanogenesis}
\alias{methanogenesis}
\title{Steps through DIC consumption during a methanogenesis reaction}
\usage{
methanogenesis(
  CH4.initial,
  K.CH4 = 0.00112896948941469,
  H2.initial,
  K.H2 = 0.000666251556662516,
  DIC.initial,
  pH.initial,
  K.CO2 = 0.023464592,
  K.CO2HCO3 = 5.223196e-07,
  K.HCO3CO3 = 6.01886e-11,
  standard.gibbs,
  temperature,
  VolumeSolution,
  VolumeHeadspace,
  delta.DIC
)
}
\arguments{
\item{CH4.initial}{Concentration of initial dissolved CH4.}

\item{K.CH4}{Henry's constant for CH4.}

\item{H2.initial}{Concentration of initial dissolved H2.}

\item{K.H2}{Henry's constant for H2.}
}
\value{
Data frame for reaction results.
}
\description{
`methanogenesis()` calculates CH4 produced, H2 consumed, CO2 consumed, and Gibbs free energy changes as dissolved inorganic carbon is consumed.
}
\examples{
methanogenesis(CH4.initial = 1e-6,H2.initial = 5e-4,DIC.initial = 3.2e-3,pH.initial = 7.5,standard.gibbs = -191359.46584,temperature = 273.15+40,VolumeSolution = 80e-3,VolumeHeadspace = 20e-3,delta.DIC = 0.0001)
}
