% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methanogenesis.R
\name{methanogenesis}
\alias{methanogenesis}
\title{Steps through DIC consumption during a methanogenesis reaction}
\usage{
methanogenesis(
  CH4.initial,
  K.CH4 = 0.00112896948941469,
  H2.initial,
  K.H2 = 0.000666251556662516,
  DIC.initial,
  pH.initial,
  K.CO2 = 0.023464592,
  standard.gibbs = -191359.46584,
  temperature,
  VolumeSolution,
  VolumeHeadspace,
  K.CO2HCO3 = 5.223196e-07,
  K.HCO3CO3 = 6.01886e-11,
  delta.DIC = 1e-04,
  biomass.yield = 2.4,
  carbon.fraction = 0.44
)
}
\arguments{
\item{CH4.initial}{Concentration of initial dissolved CH4, in molarity.}

\item{K.CH4}{Henry's constant for CH4.}

\item{H2.initial}{Concentration of initial dissolved H2, in molarity.}

\item{K.H2}{Henry's constant for H2.}

\item{DIC.initial}{Concentration of initial dissolved inorganic carbon, in molarity.}

\item{pH.initial}{initial pH.}

\item{K.CO2}{Henry's constant for CH4.}

\item{standard.gibbs}{Standard Gibbs free energy for the reaction,}

\item{temperature}{Temperature of the system, in Kelvin.}

\item{VolumeSolution}{Volume of liquid in the closed system, in liters.}

\item{VolumeHeadspace}{Volume of gaseous headspace in the closed system, in liters.}

\item{K.CO2HCO3}{Equilibrium constant for the dissociation of CO2(aq) to HCO3-(aq). 5.223196e-07 by default.}

\item{delta.DIC}{step size, in millimolar. 0.1 mM by default.}

\item{biomass.yield}{mass of dry biomass produced per mol of product. 2.4 g/mol product.}

\item{carbon.fraction}{w/w percent C of biomass, expressed as a decimal. 0.44 by default.}

\item{KHCO3CO3}{Equilibrium constant for the dissociation of HCO3- (aq) to CO3-- (aq). 6.01886e-11 by default.}
}
\value{
A data frame of the model results
}
\description{
`methanogenesis()` calculates CH4 produced, H2 consumed, CO2 consumed, and Gibbs free energy changes as dissolved inorganic carbon is consumed.
}
\details{
`init()` sets up the initial environment to be used by the methanogenesis model.
}
\examples{
methanogenesis(CH4.initial = 1e-6,H2.initial = 5e-4,DIC.initial = 3.2e-3,pH.initial = 7.5,standard.gibbs = -191359.46584,temperature = 273.15+40,VolumeSolution = 80e-3,VolumeHeadspace = 20e-3,delta.DIC = 0.0001)

}
